requirements.txt
Django==3.1.4
django-allauth==0.44.0
django-rest-auth==0.9.5
djangorestframework==3.12.2